Add files via upload
this is a signup page, that is similar to my previous signup page in appearance, in this page the image and sign up form are fit to the webpage and also when you minimize your webpage. the image disappears leaving only the textbox used to gather information at the center of the page.